Fix some bugs in printer job data format and in label dimensions calculation
Add line borders for label_conf

+++ printer/PrinterJobData.h
@@ -21,7 +21,7 @@ private:
     bool high_quality;
     bool print_600dpi;
 
-    uint8_t margin_amount;
+    uint16_t margin_amount;
 
 public:
     PrinterJobData() noexcept;
